Review related clean up (CORE-684)
Tries to make some tests more stable and independent, refactors tests in
one module to allow running in parallel to reduce build time.  Adds
automatic retries onto one modules surefire configuration to try and
avoid flakiness observed by reviewers.
