+++ jena-fuseki-kafka-module/pom.xml
@@ -212,6 +212,7 @@
                 <include>**/Test*.java</include>
                 <include>**/DockerTest*.java</include>
               </includes>
+              <rerunFailingTestsCount>1</rerunFailingTestsCount>
             </configuration>
           </plugin>
