+++ jena-fuseki-kafka-module/src/test/java/org/apache/jena/fuseki/kafka/TestFKRegistry.java
@@ -1,8 +1,10 @@
 package org.apache.jena.fuseki.kafka;
 
 import io.telicent.smart.cache.sources.kafka.KafkaTestCluster;
+import org.apache.jena.fuseki.system.FusekiLogging;
 import org.apache.jena.kafka.JenaKafkaException;
 import org.apache.jena.kafka.KConnectorDesc;
+import org.apache.jena.sys.JenaSystem;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
@@ -11,6 +13,11 @@ import java.util.List;
 import static org.mockito.Mockito.mock;
 
 public class TestFKRegistry {
+
+    static {
+        JenaSystem.init();
+        FusekiLogging.markInitialized(true);
+    }
 
     @Test(expectedExceptions = JenaKafkaException.class, expectedExceptionsMessageRegExp = "Multiple connectors.*")
     public void givenMultipleConnectors_whenRegisteringForSameTopic_thenError() {
