+++ jena-kafka-connector/src/test/java/org/apache/jena/kafka/common/AbstractDockerTests.java
@@ -9,7 +9,6 @@ import io.telicent.smart.cache.sources.kafka.KafkaTestCluster;
 import io.telicent.smart.cache.sources.kafka.serializers.RdfPayloadSerializer;
 import io.telicent.smart.cache.sources.kafka.sinks.KafkaSink;
 import io.telicent.smart.cache.sources.memory.SimpleEvent;
-import org.apache.jena.dboe.base.file.Location;
 import org.apache.jena.graph.Node;
 import org.apache.jena.graph.NodeFactory;
 import org.apache.jena.kafka.KConnectorDesc;
@@ -17,34 +16,29 @@ import org.apache.jena.kafka.SysJenaKafka;
 import org.apache.jena.sparql.core.DatasetGraph;
 import org.apache.jena.sparql.core.DatasetGraphFactory;
 import org.apache.jena.sparql.core.Quad;
+import org.apache.jena.sys.JenaSystem;
 import org.apache.jena.system.Txn;
-import org.apache.jena.tdb2.TDB2Factory;
 import org.apache.kafka.common.serialization.BytesDeserializer;
 import org.apache.kafka.common.serialization.BytesSerializer;
 import org.apache.kafka.common.utils.Bytes;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
-import java.io.IOException;
-import java.nio.file.Files;
 import java.time.Duration;
-import java.util.Collections;
-import java.util.List;
-import java.util.Properties;
-import java.util.Random;
+import java.util.*;
 import java.util.stream.Stream;
 
-public class DockerTestFusekiProjector {
+public class AbstractDockerTests {
+    static {
+        JenaSystem.init();
+    }
 
     private static final int TEST_SIZE = 10_000;
-    private static int GENERATED = 0;
-
     private static final KafkaTestCluster KAFKA = new BasicKafkaTestCluster();
     private static final Random RANDOM = new Random();
+    private static int GENERATED = 0;
 
     @BeforeAll
     public static void setup() throws InterruptedException {
@@ -53,6 +47,10 @@ public class DockerTestFusekiProjector {
         Thread.sleep(1000);
 
         // Generate test events
+        List<Node> predicates = new ArrayList<>();
+        for (int i = 0; i < 20; i++) {
+            predicates.add(NodeFactory.createURI("https://example.com/predicate/" + i));
+        }
         try (KafkaSink<Bytes, RdfPayload> sink = KafkaSink.<Bytes, RdfPayload>create()
                                                           .bootstrapServers(KAFKA.getBootstrapServers())
                                                           .producerConfig(KAFKA.getClientProperties())
@@ -60,7 +58,7 @@ public class DockerTestFusekiProjector {
                                                           .keySerializer(BytesSerializer.class)
                                                           .valueSerializer(RdfPayloadSerializer.class)
                                                           .build()) {
-            Node predicate = NodeFactory.createURI("https://example.org/predicate");
+            Node predicate = predicates.get(RANDOM.nextInt(predicates.size()));
             for (int i = 0; i < TEST_SIZE; i++) {
                 DatasetGraph dsg = DatasetGraphFactory.create();
                 int graphSize = RANDOM.nextInt(1, 1_000);
@@ -79,27 +77,6 @@ public class DockerTestFusekiProjector {
     @AfterAll
     public static void teardown() {
         KAFKA.teardown();
-    }
-
-    private void verifyProjection(EventSource<Bytes, RdfPayload> source, FusekiProjector projector, DatasetGraph dsg) {
-        long count = 0;
-        try (FusekiSink sink = FusekiSink.builder().dataset(dsg).build()) {
-            while (!source.isExhausted()) {
-                Event<Bytes, RdfPayload> event = source.poll(Duration.ofSeconds(5));
-                if (event != null) {
-                    projector.project(event, sink);
-                    count++;
-                }
-
-                if (count == TEST_SIZE) {
-                    Assertions.assertEquals(0L, source.remaining());
-                    break;
-                }
-            }
-
-            // Then
-            Assertions.assertEquals(TEST_SIZE, count);
-        }
     }
 
     private static KafkaRdfPayloadSource<Bytes> prepareSource(int batchSize, KConnectorDesc conn,
@@ -126,11 +103,36 @@ public class DockerTestFusekiProjector {
         return props;
     }
 
-    private void verifyProjection(int batchSize, DatasetGraph dsg, String consumerGroupBaseId) {
+    public static Stream<Arguments> batchSizes() {
+        return Stream.of(Arguments.of(500), Arguments.of(5000));
+    }
+
+    private void verifyProjection(EventSource<Bytes, RdfPayload> source, FusekiProjector projector, DatasetGraph dsg) {
+        long count = 0;
+        try (FusekiSink sink = FusekiSink.builder().dataset(dsg).build()) {
+            while (!source.isExhausted()) {
+                Event<Bytes, RdfPayload> event = source.poll(Duration.ofSeconds(5));
+                if (event != null) {
+                    projector.project(event, sink);
+                    count++;
+                }
+
+                if (count == TEST_SIZE) {
+                    Assertions.assertEquals(0L, source.remaining());
+                    break;
+                }
+            }
+
+            // Then
+            Assertions.assertEquals(TEST_SIZE, count);
+        }
+    }
+
+    protected void verifyProjection(int batchSize, DatasetGraph dsg, String consumerGroupBaseId) {
         // Given
-        Properties props = prepareProperties();
-        KConnectorDesc conn = prepareConnector(props);
-        KafkaRdfPayloadSource<Bytes> source = prepareSource(batchSize, conn, consumerGroupBaseId);
+        Properties props = AbstractDockerTests.prepareProperties();
+        KConnectorDesc conn = AbstractDockerTests.prepareConnector(props);
+        KafkaRdfPayloadSource<Bytes> source = AbstractDockerTests.prepareSource(batchSize, conn, consumerGroupBaseId);
         FusekiProjector projector = TestFusekiProjector.buildProjector(conn, source, dsg, batchSize);
 
         // When
@@ -138,31 +140,5 @@ public class DockerTestFusekiProjector {
 
         // Then
         Assertions.assertEquals(GENERATED, Txn.calculateRead(dsg, () -> dsg.stream().count()));
-    }
-
-    public static Stream<Arguments> batchSizes() {
-        return Stream.of(Arguments.of(500), Arguments.of(1000), Arguments.of(5000));
-    }
-
-    @MethodSource("batchSizes")
-    @ParameterizedTest
-    public void givenFusekiProjector_whenProjectingFromKafkaToInMemoryTransactionalDataset_thenDataIsLoaded(
-            int batchSize) {
-        // Given
-        DatasetGraph dsg = DatasetGraphFactory.createTxnMem();
-
-        // When and Then
-        verifyProjection(batchSize, dsg, "fuseki-kafka-projection-mem-transactional-");
-    }
-
-    @MethodSource("batchSizes")
-    @ParameterizedTest
-    public void givenFusekiProjector_whenProjectingFromKafkaToTdb2Dataset_thenDataIsLoaded(int batchSize) throws IOException {
-        // Given
-        Location location = Location.create(Files.createTempDirectory("fuseki-kafka"));
-        DatasetGraph dsg = TDB2Factory.connectDataset(location).asDatasetGraph();
-
-        // When and Then
-        verifyProjection(batchSize, dsg, "fuseki-kafka-projection-tdb2-");
     }
 }
