+++ pom.xml
@@ -62,6 +62,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.build.outputTimestamp>2025-01-24T09:38:42Z</project.build.outputTimestamp>
         <coverage.minimum>0.8</coverage.minimum>
+        <test.maxForks>2</test.maxForks>
 
         <!-- Maven Plugin versions -->
         <plugin.clean>3.4.0</plugin.clean>
